+++ ctrtool/exheader.h
@@ -154,8 +154,8 @@ typedef struct
 	u8 hash[32];
 	u8 counter[16];
 	u8 key[16];
-	u32 offset;
-	u32 size;
+	u64 offset;
+	u64 size;
 	exheader_header header;
 
 	exheader_arm11systemlocalcaps_deserialised system_local_caps;
@@ -181,8 +181,8 @@ typedef struct
 
 void exheader_init(exheader_context* ctx);
 void exheader_set_file(exheader_context* ctx, FILE* file);
-void exheader_set_offset(exheader_context* ctx, u32 offset);
-void exheader_set_size(exheader_context* ctx, u32 size);
+void exheader_set_offset(exheader_context* ctx, u64 offset);
+void exheader_set_size(exheader_context* ctx, u64 size);
 void exheader_set_partitionid(exheader_context* ctx, u8 partitionid[8]);
 void exheader_set_counter(exheader_context* ctx, u8 counter[16]);
 void exheader_set_programid(exheader_context* ctx, u8 programid[8]);
